Cleanup, reorganization and feature extensions for mrb commonality.
Substantial cleanup and reorganization.

More functionality in common with mrb: parse_deps.pm is intended to be synced between mrb. Specifically:

  1. Some functions have been moved or created from code from elsewhere.

  2. Some functions have been made more flexibility to accommodate mrb requirements.

  3. Version handling is more flexible, and handles alpha, beta, patch and pre versions.

The product table format has evolved, including a '<format=#>' indicator. Specifiying '<format=2>' allows the use of '-' in the product list for its intended purpose (unconditional inclusion). The meaning of '-nq-' changes to require "no qualifier", as it does in the qualifier table.